datafixture added, enums replaced with entity

+++ Airline/Domain/Entities/AircraftModel.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Domain.Enums;
 
 namespace Domain.Entities;
+
 public class AircraftModel
 {
     public required int Id { get; set; }
@@ -14,7 +14,7 @@ public class AircraftModel
 
     public required AircraftFamily ModelFamily { get; set; }
 
-    public required float FlightRange { get; set; }
+    public required int FlightRange { get; set; }
 
     public required int PassengerCapacity { get; set; }
